requirements.txt
streamlit
nltk
scikit-learn
sentence-transformers
torch
google-generativeai